P2/P2_Falta2YMemoria/Ejer_1/requirements.txt
# Core dependencies
neat-python>=0.92
numpy>=1.21.0
gymnasium>=0.28.0

# Visualization
matplotlib>=3.5.0
seaborn>=0.12.0
graphviz>=0.20.0


robobopy
robobosim
